add a new project hoback
hoback is a river in wyoming

Wikipedia says:

The Hoback River, once called the Fall River, is an approximately 55-mile (89 km)-long tributary of the Snake River in the U.S. state of Wyoming. It heads in the northern Wyoming Range of Wyoming and flows north and then east through the Bridger-Teton National Forest.

This project is like a river.
It is an experimental learning experience.